requirements.txt
Django == 3.2.6
psycopg2 == 2.9.1
Pillow == 8.3.1
djangorestframework == 3.12.4
httpie == 2.4.0
pandas == 1.3.3
python-dotenv == 0.19.0